+++ src/components/aidreq-list-Item/ui/top/indexCss.ts
@@ -27,39 +27,6 @@ export const Text = styled.p`
   color: #535353;
   height: 28px;
   overflow: hidden;
-
-  h1 {
-    font-size: 2em;
-    font-weight: bold;
-    margin-bottom: 1em;
-  }
-
-  h2 {
-    font-size: 1.5em;
-    font-weight: bold;
-    margin-bottom: 0.8em;
-  }
-
-  h3 {
-    font-size: 1.17em;
-    font-weight: bold;
-    margin-bottom: 0.6em;
-  }
-
-  ul,
-  ol {
-    margin: 1em 0;
-    padding-left: 2em;
-  }
-
-  li {
-    list-style: disc;
-    margin: 0.5em 0;
-  }
-
-  strong {
-    font-weight: bold;
-  }
 `;
 
 export const Center = styled.p`
