--- src/components/content-viewer/indexCss.ts
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  width: 100%;
-
-  img {
-    max-width: 100%;
-  }
-`;
