+++ src/components/sanitized-content/indexCss.ts
@@ -1,0 +1,36 @@
+import styled from 'styled-components';
+
+export const SanitizedContentWrapper = styled.div`
+  h1 {
+    font-size: 2em;
+    font-weight: bold;
+    margin-bottom: 1em;
+  }
+
+  h2 {
+    font-size: 1.5em;
+    font-weight: bold;
+    margin-bottom: 0.8em;
+  }
+
+  h3 {
+    font-size: 1.17em;
+    font-weight: bold;
+    margin-bottom: 0.6em;
+  }
+
+  ul,
+  ol {
+    margin: 1em 0;
+    padding-left: 2em;
+  }
+
+  li {
+    list-style: disc;
+    margin: 0.5em 0;
+  }
+
+  strong {
+    font-weight: bold;
+  }
+`;
