+++ src/features/personal-my-page-review-list/_component/ReviewBox.tsx
@@ -1,3 +1,5 @@
+import sanitizeHtml from 'sanitize-html';
+import parse from 'html-react-parser';
 import useDateFormat from '@/shared/hooks/useDateFormat';
 import ReviewBoxCss from './ReviewBoxCss';
 
@@ -10,6 +12,12 @@ interface ReviewBoxProps {
 
 const ReviewBox: React.FC<ReviewBoxProps> = ({ orgName, createdDate, reviewText, onClick }) => {
   const { formatDateTime } = useDateFormat();
+
+  const sanitizedContent = sanitizeHtml(reviewText, {
+    allowedTags: ['p'], // p 태그만 허용하기
+    allowedAttributes: {} // 어떤 요소도 허용 x
+  });
+
   return (
     <ReviewBoxCss onClick={onClick}>
       <div className="infoWrap">
@@ -17,7 +25,7 @@ const ReviewBox: React.FC<ReviewBoxProps> = ({ orgName, createdDate, reviewText,
         <i>{formatDateTime(createdDate)}</i>
       </div>
 
-      <p>{reviewText}</p>
+      <p>{parse(sanitizedContent)}</p>
     </ReviewBoxCss>
   );
 };
